improve logic add constant change to double

+++ features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/data/model/ratesestimate/P2RatesEstimate.kt
@@ -36,7 +36,7 @@ data class P2RatesEstimateData(
 
         @SerializedName("cheapestShippingPrice")
         @Expose
-        val cheapestShippingPrice: Long = 0L,
+        val cheapestShippingPrice: Double = 0.0,
 
         @SerializedName("destination")
         @Expose
